Shifting back bulk subscription to runtime file
